Feat: add some errors messages

+++ src/profile/forms.py
@@ -152,20 +152,44 @@ class RatingResponseForm(forms.ModelForm):
         fields = ['response_text']
 
 class FreelancerRegisterForm(UserCreationForm):
-    identification = forms.CharField(max_length=20, required=True)
-    phone = forms.CharField(max_length=15, required=True)
-    photo = forms.ImageField(required=False)
+    identification = forms.CharField(max_length=20, required=True, label='ID de Identificación')
+    phone = forms.CharField(max_length=15, required=True, label='Teléfono')
+    photo = forms.ImageField(required=False, label='Foto de Perfil')
 
     class Meta:
         model = User
         fields = ['username', 'email', 'password1', 'password2']
+        labels = {
+            'username': 'Nombre de Usuario',
+            'email': 'Correo Electrónico',
+            'password1': 'Contraseña',
+            'password2': 'Confirmar Contraseña',
+        }
+    
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields['password1'].label = 'Contraseña'
+        self.fields['password2'].label = 'Confirmar Contraseña'
+        self.order_fields(['username', 'identification', 'phone', 'email', 'password1', 'password2', 'photo'])
+
+        
+    def clean_email(self):
+        email = self.cleaned_data.get('email')
+        if User.objects.filter(email=email).exists():
+            raise forms.ValidationError('Este correo electrónico ya está registrado. Por favor, usa otro.')
+        return email
+
+    def clean_identification(self):
+        identification = self.cleaned_data.get('identification')
+        if FreelancerProfile.objects.filter(identification=identification).exists():
+            raise forms.ValidationError('Este ID ya está registrado. Por favor, usa otro.')
+        return identification
 
     def save(self, commit=True):
         user = super().save(commit=False)
         if commit:
             user.save()
             try:
-                # Intentamos crear el perfil del freelancer
                 FreelancerProfile.objects.create(
                     user=user,
                     identification=self.cleaned_data['identification'],
@@ -173,26 +197,46 @@ class FreelancerRegisterForm(UserCreationForm):
                     photo=self.cleaned_data.get('photo')
                 )
             except IntegrityError as e:
-                # Si ocurre un error de integridad, agregamos el error al formulario
-                self.add_error('identification', 'Este ID ya está registrado. Por favor, usa otro.')
-                user.delete()  # Borra el usuario para evitar usuarios sin perfiles
-                raise e  # Lanza la excepción para ser manejada en la vista
+                if 'unique constraint' in str(e).lower() and 'identification' in str(e).lower():
+                    self.add_error('identification', 'Este ID ya está registrado. Por favor, usa otro.')
+                    user.delete() 
+                return user  
         return user
 
 
+
 class CompanyRegisterForm(UserCreationForm):
-    company_name = forms.CharField(max_length=100, required=True)
-    nit = forms.CharField(max_length=20, required=True)
-    business_type = forms.CharField(max_length=50, required=True)
-    country = forms.CharField(max_length=50, required=True)
-    business_vertical = forms.CharField(max_length=50, required=True)
-    address = forms.CharField(max_length=150, required=True)
-    legal_representative = forms.CharField(max_length=100, required=True)
-    phone = forms.CharField(max_length=15, required=True)
+    company_name = forms.CharField(max_length=100, required=True, label='Nombre de la Empresa')
+    nit = forms.CharField(max_length=20, required=True, label='NIT')
+    business_type = forms.CharField(max_length=50, required=True, label='Tipo de Empresa')
+    country = forms.CharField(max_length=50, required=True, label='País')
+    business_vertical = forms.CharField(max_length=50, required=True, label='Vertical de Negocio')
+    address = forms.CharField(max_length=150, required=True, label='Dirección')
+    legal_representative = forms.CharField(max_length=100, required=True, label='Representante Legal')
+    phone = forms.CharField(max_length=15, required=True, label='Teléfono')
+    photo = forms.ImageField(required=False, label='Imagen de la Empresa')
 
     class Meta:
         model = User
         fields = ['username', 'email', 'password1', 'password2']
+        labels = {
+            'username': 'Nombre de Usuario',
+            'email': 'Correo Electrónico',
+            'password1': 'Contraseña',
+            'password2': 'Confirmar Contraseña',
+        }
+        
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields['password1'].label = 'Contraseña'
+        self.fields['password2'].label = 'Confirmar Contraseña'
+        self.order_fields(['username', 'company_name', 'nit', 'business_type', 'country', 'business_vertical', 'address', 'legal_representative', 'phone', 'email', 'password1', 'password2', 'photo'])
+        
+    def clean_email(self):
+        email = self.cleaned_data.get('email')
+        if User.objects.filter(email=email).exists():
+            raise forms.ValidationError('Este correo electrónico ya está registrado. Por favor, usa otro.')
+        return email
 
     def save(self, commit=True):
         user = super().save(commit=False)
@@ -208,14 +252,11 @@ class CompanyRegisterForm(UserCreationForm):
                     business_vertical=self.cleaned_data['business_vertical'],
                     address=self.cleaned_data['address'],
                     legal_representative=self.cleaned_data['legal_representative'],
-                    phone=self.cleaned_data['phone']
+                    phone=self.cleaned_data['phone'],
+                    photo=self.cleaned_data.get('photo')
                 )
         except IntegrityError as e:
-            # Verificamos cuál es el campo que genera el error de integridad
             if 'unique constraint' in str(e).lower() and 'nit' in str(e).lower():
                 self.add_error('nit', 'Este NIT ya está registrado. Por favor, usa otro.')
-            if 'unique constraint' in str(e).lower() and 'username' in str(e).lower():
-                self.add_error('username', 'Este nombre de usuario ya está registrado. Elige otro.')
-            user.delete()  # Borra el usuario para evitar usuarios sin perfiles
-            raise e  # Lanza la excepción para ser manejada en la vista
+            raise e  
         return user
